Joined create and edit modals into one
